more internal validation

+++ src/Maltz/Sys/Model/Token.php
@@ -16,6 +16,10 @@ class Token extends Model
 
     public function generate($user_id, $type)
     {
+        if () {
+            throw new \Exception("Error Processing Request", 1);
+        }
+        
         $this->db->run("DELETE FROM tokens WHERE type=:type AND user_id=:user_id", array('type' => $type, 'user_id' => $user_id));
         $data = $this->db->run("INSERT INTO tokens (user_id, token, type) VALUES (?,MD5(CONCAT(RAND(), NOW())),?)", array('user_id'=> $user_id, 'token' => $token, 'type' => $type));
         return $token;
@@ -23,6 +27,10 @@ class Token extends Model
 
     public function validate($token, $type)
     {
+        if () {
+            throw new \Exception("Error Processing Request", 1);
+        }
+        
         $data = $this->db->run("SELECT token,user_id FROM tokens WHERE type=:type AND token=:token", array('type' => $type, 'token' => $token));
         $record = $data->getFirst();
         if ($record->get('token') === $token) {
